// src/app/models/tickerPrice.ts
export class tickerPrice {
    prevClose: number;
    mid: number;
    lastSaleTimestamp: string;
    open: number;
    askPrice: number;
    low: number;
    ticker: string;
    timestamp: string;
    lastSize: number;
    tngoLast: number;
    last: number;
    high: number;
    askSize: number;
    quoteTimestamp: string;
    bidPrice: number;
    bidSize: number;
    volume: bigint;
}